[Upgrade controllers] user id dependence

+++ src/main/java/io/github/denkoch/mycosts/controller/CategoryController.java
@@ -1,13 +1,14 @@
 package io.github.denkoch.mycosts.controller;
 
 import io.github.denkoch.mycosts.entities.Category;
+import io.github.denkoch.mycosts.entities.Payment;
 import io.github.denkoch.mycosts.service.CategoryService;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
 @Controller
-@RequestMapping(value = "/users/{id}")
+@RequestMapping(value = "/users/{userId}")
 public class CategoryController {
 
     private final CategoryService categoryService;
@@ -16,31 +17,36 @@ public class CategoryController {
         this.categoryService = categoryService;
     }
 
-    @ModelAttribute("id")
-    public Integer populateId(@PathVariable Integer id) {
-        return id;
+    @ModelAttribute("userId")
+    public Integer populateId(@PathVariable Integer userId) {
+        return userId;
     }
 
     @RequestMapping(value = "/categories", method = RequestMethod.GET)
-    public String getCategory(Model model) {
+    public String getCategories(@PathVariable Integer userId, Model model) {
 
-        model.addAttribute("categories", categoryService.getCategories());
-        model.addAttribute("category", new Category(categoryService.getMaxId() + 1));
+        model.addAttribute("categories", categoryService.getCategories(userId));
+        model.addAttribute("category", new Category(categoryService.createId(), userId));
         return "categories";
     }
 
     @RequestMapping(value = "/categories", method = RequestMethod.POST, params = {"add"})
-    public String postPayments(@ModelAttribute("category") Category category) {
+    public String postPayments(@PathVariable Integer userId,
+                               @ModelAttribute("category") Category category) {
 
-        categoryService.addCategory(category);
+        categoryService.addCategory(userId, category);
 
         return "redirect:categories";
     }
 
     @RequestMapping(value = "/categories", method = RequestMethod.POST, params = {"delete"})
-    public String postPayments(@RequestParam("category-id") Integer categoryId) {
+    public String postPayments(@PathVariable Integer userId,
+                               @RequestParam("category-id") Integer categoryId) {
 
-        categoryService.deleteCategory(categoryId);
+        Category temp = categoryService.getCategory(categoryId);
+        if (temp != null && temp.getUserId().equals(userId)) {
+            categoryService.deleteCategory(categoryId);
+        }
 
         return "redirect:categories";
     }
